Rough Camera and some refactoring
Added a CameraController with Pan, Zoom, Roatate. Currently the camera is roatating incorrectly (needs to rotate around the camera's center)
Created an InputController to handle input (such as camera movement)
Added PlayerController to handle dealing with the player (moved from Loader), need to figure out how to set the player's current location.
Added a static GameMaster that will collect the various objects that will need to be delt with as well as a GameBase that MonoBehavior's can inherit from in order to have easy access to GameMaster (GameBase.Game)